rewrite the item filters code
Now item_filters is a feed specific setting, that accepts a comma
separated list of filters to apply to the items in the feed.

rss2maildir obtains the filters from the file filters.py in the config
directory (~/.config/rss2maildir by default) every filter is a function

def my_filter(item):
   # bla bla

   return item

It accepts a single parameter item, and must return the modified item,
or None if the item is to be ignored.