Get rid of item database
Now we store md5 sum of messages into the item filename. This way, we
can check for seen messages directly in the maildir structure.

This is more robust. For instance, with the old database scheme, deleted
items would remain in the database forever.

Also, now it is much easier to parallelize!